# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillMatch AI Services API",
    "description": "Production AI Services API for SkillMatch. The canonical AI-Serv5 CV extraction pipeline powers CV, matching, interview, and review features.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/cv/extract-file": {
      "post": {
        "tags": [
          "CV Profile Extraction"
        ],
        "summary": "Extract Structured Profile from CV File (Direct)",
        "description": "Upload a digital CV/Resume file (PDF, DOCX, TXT, or Markdown). Processes the file without blocking the FastAPI event loop via background thread pool.",
        "operationId": "extract_cv_file_api_v1_cv_extract_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_cv_file_api_v1_cv_extract_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully extracted and normalized candidate profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsupported file format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "File size exceeds allowed limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error in file input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "CV extraction service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/cv/extract-file-async": {
      "post": {
        "tags": [
          "CV Profile Extraction"
        ],
        "summary": "Submit CV File for Async Background Extraction",
        "description": "Enqueues a CV extraction task as an async background job (as required by README.md:66). Returns a pollable job_id.",
        "operationId": "extract_cv_file_async_api_v1_cv_extract_file_async_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_cv_file_async_api_v1_cv_extract_file_async_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Job successfully queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CVJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or unsupported file format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "File size exceeds allowed limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/cv/jobs/{job_id}": {
      "get": {
        "tags": [
          "CV Profile Extraction"
        ],
        "summary": "Poll Background CV Extraction Job Status",
        "description": "Returns the status of a background CV extraction job (queued, processing, completed, failed) and candidate profile upon completion.",
        "operationId": "get_job_status_api_v1_cv_jobs__job_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current status of the job",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CVJobResponse"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cv/extract-text": {
      "post": {
        "tags": [
          "CV Profile Extraction"
        ],
        "summary": "Extract Structured Profile from Raw CV Text",
        "description": "Extracts a candidate profile from raw CV text using the canonical AI-Serv5 pipeline.",
        "operationId": "extract_cv_text_api_v1_cv_extract_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractTextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully extracted candidate profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candidate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input text payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "CV extraction service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_extract_cv_file_api_v1_cv_extract_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "CV file to upload (PDF, DOCX, TXT, or Markdown)"
          },
          "candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Id",
            "description": "Optional custom candidate ID (e.g. cand_001)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_extract_cv_file_api_v1_cv_extract_file_post"
      },
      "Body_extract_cv_file_async_api_v1_cv_extract_file_async_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "CV file to upload (PDF, DOCX, TXT, or Markdown)"
          },
          "candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Id",
            "description": "Optional custom candidate ID (e.g. cand_001)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_extract_cv_file_async_api_v1_cv_extract_file_async_post"
      },
...
      "CVJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "Unique background job identifier"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "description": "Current status of the background extraction job",
            "default": "queued"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Candidate"
              },
              {
                "type": "null"
              }
            ],
            "description": "Extracted Candidate profile if status is completed"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Error message if status is failed"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code",
            "description": "Standardized error code if status is failed"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "CVJobResponse"
      },
      "Candidate": {
        "properties": {
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id",
            "description": "Unique candidate identifier, e.g. cand_001"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfile",
            "description": "User entity fields: name, email"
          },
          "candidate_profile": {
            "$ref": "#/components/schemas/CandidateProfile",
            "description": "Candidate profile fields"
          },
          "educations": {
            "items": {
              "$ref": "#/components/schemas/EducationItem"
            },
            "type": "array",
            "title": "Educations",
            "description": "Education records"
          },
          "experiences": {
            "items": {
              "$ref": "#/components/schemas/ExperienceItem"
            },
            "type": "array",
            "title": "Experiences",
            "description": "Work experience records"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectItem"
            },
            "type": "array",
            "title": "Projects",
            "description": "Academic, personal, or professional projects"
          },
          "certificates": {
            "items": {
              "$ref": "#/components/schemas/CertificateItem"
            },
            "type": "array",
            "title": "Certificates",
            "description": "Certificates & courses"
          },
          "languages": {
            "items": {
              "$ref": "#/components/schemas/LanguageItem"
            },
            "type": "array",
            "title": "Languages",
            "description": "Languages spoken and proficiency"
          },
          "candidate_skills": {
            "items": {
              "$ref": "#/components/schemas/CandidateSkill-Output"
            },
            "type": "array",
            "title": "Candidate Skills",
            "description": "Extracted & normalized skills with evidence and confidence"
          },
          "target_roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Roles",
            "description": "Target job roles / career tracks"
          },
          "preferences": {
            "$ref": "#/components/schemas/CandidatePreferences",
            "description": "Career & job preferences"
          }
        },
        "type": "object",
        "title": "Candidate"
      },
      "CandidatePreferences": {
        "properties": {
          "employment_type": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Employment Type",
            "description": "Preferred employment types"
          },
          "work_mode": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Work Mode",
            "description": "Preferred work modes (remote, hybrid, onsite)"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locations",
            "description": "Preferred geographic locations"
          },
          "industries": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Industries",
            "description": "Target industries"
          }
        },
        "type": "object",
        "title": "CandidatePreferences"
      },
      "CandidateProfile": {
        "properties": {
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline",
            "description": "Professional headline if explicitly stated in CV"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio",
            "description": "Professional summary/bio if explicitly stated in CV"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone",
            "description": "Phone number"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Candidate's personal location"
          },
          "linkedin_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linkedin Url",
            "description": "LinkedIn profile URL if present in CV"
          },
          "github_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Url",
            "description": "Personal GitHub profile URL if present in CV"
          },
          "portfolio_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Url",
            "description": "Portfolio or personal website URL if present in CV"
          }
        },
        "type": "object",
        "title": "CandidateProfile"
      },
...
      "ExtractTextRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Raw text of the CV to extract"
          },
          "candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Id",
            "description": "Optional custom candidate ID"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ExtractTextRequest"
      },
      "ExtractionErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Description of the error that occurred"
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Standardized error code",
            "example": "UNSUPPORTED_FILE_TYPE"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error_code"
        ],
        "title": "ExtractionErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "processing",
          "completed",
          "failed"
        ],
        "title": "JobStatus"
      },
      "LanguageItem": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Language name, e.g. English, Arabic"
          },
          "proficiency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proficiency",
            "description": "Proficiency level if explicitly stated in CV (e.g. Native, Fluent, Intermediate)"
          }
        },
        "type": "object",
        "required": [
          "language"
        ],
        "title": "LanguageItem"
      },
...
      "UserProfile": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Candidate full name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Candidate email address"
          }
        },
        "type": "object",
        "title": "UserProfile"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}